+++ Source/Core/Addons.cpp
@@ -1,7 +1,9 @@
 #include "Addons.h"
 
+#include "Logger.h"
+
 void MethaneVersion() {
-	std::println("Using Methane SDL OpenGL Framework v2.07");
+	Logger::Log("Using Methane SDL OpenGL Framework v2.07");
 }