+++ Source/Programs/Files.cpp
@@ -1,7 +1,6 @@
 #include "Files.h"
 
 #include <fstream>
-#include <print>
 #include <filesystem>
 
 std::string BreakLines(const std::string& input, std::size_t every) {